# docs/openapi/repository-api-v1.json
{
  "components": {
    "schemas": {
...
      "ErrorBody": {
        "description": "Error detail object for the public API error envelope.",
        "properties": {
          "code": {
            "description": "Stable machine-readable error code.",
            "title": "Code",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional structured metadata to help clients debug the error.",
            "title": "Details"
          },
          "message": {
            "description": "Human-readable summary of the failure.",
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorEnvelope": {
        "description": "Standardized error envelope shared across HTTP endpoints.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody",
            "description": "Normalized error payload."
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
      "ExactSkillCoordinateRequest": {
        "additionalProperties": false,
        "description": "Exact immutable skill-version coordinate used by fetch and resolution reads.",
        "properties": {
          "skill_id": {
            "description": "Stable identifier of the requested skill.",
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9](?:[A-Za-z0-9._-]{0,127})$",
            "title": "Skill Id",
            "type": "string"
          },
          "version": {
            "description": "Exact immutable semantic version of the requested skill.",
            "pattern": "^(0|[1-9]\\d*)\\.(0|[1-9]\\d*)\\.(0|[1-9]\\d*)(?:-([0-9A-Za-z-]+(?:\\.[0-9A-Za-z-]+)*))?(?:\\+([0-9A-Za-z-]+(?:\\.[0-9A-Za-z-]+)*))?$",
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "skill_id",
          "version"
        ],
        "title": "ExactSkillCoordinateRequest",
        "type": "object"
      },
...
      "RelatedSkillVersionSummaryResponse": {
        "description": "Compact exact target summary for direct relationships with exact selectors.",
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional human-readable summary.",
            "title": "Description"
          },
          "name": {
            "description": "Human-readable skill name.",
            "title": "Name",
            "type": "string"
          },
          "published_at": {
            "description": "UTC timestamp when the related version was published.",
            "format": "date-time",
            "title": "Published At",
            "type": "string"
          },
          "skill_id": {
            "description": "Stable identifier of the related skill.",
            "title": "Skill Id",
            "type": "string"
          },
          "tags": {
            "description": "Searchable tags attached to the related version.",
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          },
          "version": {
            "description": "Exact immutable semantic version of the related skill.",
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "skill_id",
          "version",
          "name",
          "description",
          "tags",
          "published_at"
        ],
        "title": "RelatedSkillVersionSummaryResponse",
        "type": "object"
      },
...
      "RelationshipSelectorResponse": {
        "additionalProperties": false,
        "description": "Authored relationship selector preserved from the immutable manifest.",
        "properties": {
          "markers": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Execution markers preserved exactly as authored.",
            "title": "Markers"
          },
          "optional": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional dependency marker preserved as authored.",
            "title": "Optional"
          },
          "skill_id": {
            "description": "Stable identifier of the related skill.",
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9](?:[A-Za-z0-9._-]{0,127})$",
            "title": "Skill Id",
            "type": "string"
          },
          "version": {
            "anyOf": [
              {
                "pattern": "^(0|[1-9]\\d*)\\.(0|[1-9]\\d*)\\.(0|[1-9]\\d*)(?:-([0-9A-Za-z-]+(?:\\.[0-9A-Za-z-]+)*))?(?:\\+([0-9A-Za-z-]+(?:\\.[0-9A-Za-z-]+)*))?$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Exact related version when the relationship targets one immutable version.",
            "title": "Version"
          },
          "version_constraint": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Authored dependency constraint when the relationship is version-ranged.",
            "title": "Version Constraint"
          }
        },
        "required": [
          "skill_id"
        ],
        "title": "RelationshipSelectorResponse",
        "type": "object"
      },
...
      "SkillRelationshipBatchItemResponse": {
        "description": "One ordered direct relationship lookup result.",
        "properties": {
          "coordinate": {
            "$ref": "#/components/schemas/ExactSkillCoordinateRequest",
            "description": "Exact immutable source coordinate as requested by the client."
          },
          "relationships": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/SkillRelationshipEdgeResponse"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Direct authored relationships for the source version when it exists.",
            "title": "Relationships"
          },
          "status": {
            "description": "Whether the requested source coordinate was found.",
            "enum": [
              "found",
              "not_found"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "coordinate"
        ],
        "title": "SkillRelationshipBatchItemResponse",
        "type": "object"
      },
      "SkillRelationshipBatchRequest": {
        "additionalProperties": false,
        "description": "Ordered direct relationship query over immutable source versions.",
        "properties": {
          "coordinates": {
            "description": "Ordered immutable source coordinates to inspect.",
            "items": {
              "$ref": "#/components/schemas/ExactSkillCoordinateRequest"
            },
            "maxItems": 100,
            "minItems": 1,
            "title": "Coordinates",
            "type": "array"
          },
          "edge_types": {
            "description": "Relationship families to include. Defaults to `depends_on` and `extends`.",
            "items": {
              "enum": [
                "depends_on",
                "extends"
              ],
              "type": "string"
            },
            "title": "Edge Types",
            "type": "array"
          }
        },
        "required": [
          "coordinates"
        ],
        "title": "SkillRelationshipBatchRequest",
        "type": "object"
      },
      "SkillRelationshipBatchResponse": {
        "description": "Ordered batch response for direct immutable relationship reads.",
        "properties": {
          "results": {
            "description": "Ordered per-source relationship results preserving request order.",
            "items": {
              "$ref": "#/components/schemas/SkillRelationshipBatchItemResponse"
            },
            "title": "Results",
            "type": "array"
          }
        },
        "required": [
          "results"
        ],
        "title": "SkillRelationshipBatchResponse",
        "type": "object"
      },
      "SkillRelationshipEdgeResponse": {
        "description": "One direct authored relationship edge from an immutable source version.",
        "properties": {
          "edge_type": {
            "description": "Relationship family preserved from the manifest.",
            "enum": [
              "depends_on",
              "extends"
            ],
            "title": "Edge Type",
            "type": "string"
          },
          "selector": {
            "$ref": "#/components/schemas/RelationshipSelectorResponse",
            "description": "Authored selector preserved exactly from the immutable manifest."
          },
          "target_version": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RelatedSkillVersionSummaryResponse"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional exact target version summary when the selector points to one exact version."
          }
        },
        "required": [
          "edge_type",
          "selector"
        ],
        "title": "SkillRelationshipEdgeResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Registry-first API for immutable skill publication, indexed search candidate\nretrieval, direct relationship reads, exact version retrieval, artifact streaming,\nand version listing.\n\nThe server owns data-local registry operations only. Prompt interpretation,\nreranking, dependency solving, final selection, lock generation, and execution\nplanning remain client-owned behavior outside this API boundary.",
    "title": "aptitude-server Service",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/resolution/relationships:batch": {
      "post": {
        "description": "Return authored direct `depends_on` and `extends` relationships for exact immutable source versions. This route does not expand transitive graphs, select versions for constraints, or emit solved dependency closures.",
        "operationId": "batchGetDirectSkillRelationships",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SkillRelationshipBatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "results": [
                    {
                      "coordinate": {
                        "skill_id": "python.lint",
                        "version": "1.2.3"
                      },
                      "relationships": [
                        {
                          "edge_type": "depends_on",
                          "selector": {
                            "markers": [
                              "linux",
                              "gpu"
                            ],
                            "optional": true,
                            "skill_id": "python.base",
                            "version_constraint": ">=1.0.0,<2.0.0"
                          }
                        },
                        {
                          "edge_type": "extends",
                          "selector": {
                            "skill_id": "python.base",
                            "version": "1.0.0"
                          },
                          "target_version": {
                            "description": "Base runtime skill",
                            "name": "Python Base",
                            "published_at": "2026-03-01T08:30:00Z",
                            "skill_id": "python.base",
                            "tags": [
                              "python",
                              "runtime"
                            ],
                            "version": "1.0.0"
                          }
                        }
                      ],
                      "status": "found"
                    },
                    {
                      "coordinate": {
                        "skill_id": "python.missing",
                        "version": "9.9.9"
                      },
                      "status": "not_found"
                    }
                  ]
                },
                "schema": {
                  "$ref": "#/components/schemas/SkillRelationshipBatchResponse"
                }
              }
            },
            "description": "Ordered direct relationship results returned successfully."
          },
          "422": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "INVALID_REQUEST",
                    "details": {
                      "errors": [
                        {
                          "input": "latest",
                          "loc": [
                            "path",
                            "version"
                          ],
                          "msg": "String should match pattern",
                          "type": "string_pattern_mismatch"
                        }
                      ]
                    },
                    "message": "Request validation failed."
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The relationship query request is invalid."
          }
        },
        "summary": "Read direct immutable relationships",
        "tags": [
          "resolution"
        ]
      }
    },
...
  }
}